disk_cache: Correct SparseData length and offset type in memory cache
This CL makes SparseData related length values to size_t properly in
memory cache instead of using int64_t everywhere.

Each sparse range size must fit in size_t value as it reads from buffer.
On the other hand, the offset should be uint64_t as each size must fit
in size_t but the totalsize might exceed.

Bug: 391398191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651566
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476015}

+++ net/disk_cache/memory/mem_entry_impl.h
@@ -150,9 +150,9 @@ class NET_EXPORT_PRIVATE MemEntryImpl final
   int InternalReadData(int index, int offset, IOBuffer* buf, int buf_len);
   int InternalWriteData(int index, int offset, IOBuffer* buf, int buf_len,
                         bool truncate);
-  int InternalReadSparseData(int64_t offset, IOBuffer* buf, int buf_len);
-  int InternalWriteSparseData(int64_t offset, IOBuffer* buf, int buf_len);
-  RangeResult InternalGetAvailableRange(int64_t offset, int len);
+  int InternalReadSparseData(uint64_t offset, IOBuffer* buf, size_t buf_len);
+  int InternalWriteSparseData(uint64_t offset, IOBuffer* buf, size_t buf_len);
+  RangeResult InternalGetAvailableRange(uint64_t offset, size_t len);
 
   // Initializes the children map and sparse info. This method is only called
   // on a parent entry.
@@ -162,12 +162,12 @@ class NET_EXPORT_PRIVATE MemEntryImpl final
   // child entry or this entry itself if |offset| points to the first range.
   // If such entry does not exist and |create| is true, a new child entry is
   // created.
-  MemEntryImpl* GetChild(int64_t offset, bool create);
+  MemEntryImpl* GetChild(uint64_t offset, bool create);
 
   // Returns an interval describing what's stored in the child entry pointed to
   // by i, in global coordinates.
   // Precondition: i != children_.end();
-  net::Interval<int64_t> ChildInterval(
+  net::Interval<uint64_t> ChildInterval(
       MemEntryImpl::EntryMap::const_iterator i);
 
   // Compact vectors to try to avoid over-allocation due to exponential growth.
@@ -178,9 +178,9 @@ class NET_EXPORT_PRIVATE MemEntryImpl final
   uint32_t ref_count_ = 0;
 
   int64_t child_id_;     // The ID of a child entry.
-  int child_first_pos_ = 0;  // The position of the first byte in a child
-                             // entry. 0 here is beginning of child, not of
-                             // the entire file.
+  uint64_t child_first_pos_ = 0;  // The position of the first byte in a child
+                                  // entry. 0 here is beginning of child, not of
+                                  // the entire file.
   // Pointer to the parent entry, or nullptr if this entry is a parent entry.
   raw_ptr<MemEntryImpl> parent_;
   std::unique_ptr<EntryMap> children_;
